Apply map overlay updates to source paths
