--- App.css
@@ -1,46 +1,0 @@
-/*
- * Root shell styles only. Map, sidebar, Leaflet, and popup workflow styling
- * belongs in index.css or the owning component modules.
- */
-.app-shell-main {
-  width: 100%;
-  height: 100%;
-}
-
-.app-skip-link {
-  display: none !important;
-}
-
-.app-skip-link:focus-visible {
-  display: none !important;
-}
-
-.app-shell-loading {
-  min-height: var(--app-height, 100vh);
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  justify-content: center;
-  gap: 0.5rem;
-  text-align: center;
-  padding: 1.5rem;
-  background:
-    radial-gradient(960px 520px at 12% -8%, rgba(255, 245, 229, 0.72), transparent 56%),
-    radial-gradient(920px 540px at 92% -12%, rgba(220, 231, 238, 0.72), transparent 58%),
-    linear-gradient(165deg, #f5f5f7, #e7ebef);
-  color: #18212b;
-}
-
-.app-shell-loading h1 {
-  margin: 0;
-  font-family: 'Manrope', -apple-system, BlinkMacSystemFont, 'Avenir Next', 'Segoe UI', sans-serif;
-  font-size: clamp(1.4rem, 2.8vw, 2.25rem);
-  font-weight: 700;
-  letter-spacing: -0.02em;
-}
-
-.app-shell-loading p {
-  margin: 0;
-  font-family: 'Manrope', -apple-system, BlinkMacSystemFont, 'Avenir Next', 'Segoe UI', sans-serif;
-  color: #677381;
-}
